test(grpc): address review round 3 - shared interceptor, test hygiene
- Extract the duplicated AuthClientInterceptor into a shared package-private
  GrpcTestAuthInterceptor in the grpcw test sources; both gRPC type tests use it.
- Make TimeSeriesGrpcInsertResultTypeIT package-private (test convention).
- Move the SERVER_PLUGINS reset into a finally block so cleanup is unconditional.
