+++ grpcw/src/test/java/com/arcadedb/server/grpc/TimeSeriesGrpcInsertResultTypeIT.java
@@ -20,16 +20,10 @@ package com.arcadedb.server.grpc;
 
 import com.arcadedb.GlobalConfiguration;
 import com.arcadedb.server.BaseGraphServerTest;
-import io.grpc.CallOptions;
 import io.grpc.Channel;
-import io.grpc.ClientCall;
-import io.grpc.ClientInterceptor;
 import io.grpc.ClientInterceptors;
-import io.grpc.ForwardingClientCall;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
-import io.grpc.Metadata;
-import io.grpc.MethodDescriptor;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -45,17 +39,10 @@ import static org.assertj.core.api.Assertions.assertThat;
  * "Type with name '' was not found" on every TS insert while materialising the response, even though
  * the server had already committed the point.
  */
-public class TimeSeriesGrpcInsertResultTypeIT extends BaseGraphServerTest {
+class TimeSeriesGrpcInsertResultTypeIT extends BaseGraphServerTest {
 
   private static final int    GRPC_PORT = 50051;
   private static final String TYPE_NAME = "sensor";
-
-  private static final Metadata.Key<String> USER_HEADER     =
-      Metadata.Key.of("x-arcade-user", Metadata.ASCII_STRING_MARSHALLER);
-  private static final Metadata.Key<String> PASSWORD_HEADER =
-      Metadata.Key.of("x-arcade-password", Metadata.ASCII_STRING_MARSHALLER);
-  private static final Metadata.Key<String> DATABASE_HEADER =
-      Metadata.Key.of("x-arcade-database", Metadata.ASCII_STRING_MARSHALLER);
 
   private ManagedChannel                                  channel;
   private ArcadeDbServiceGrpc.ArcadeDbServiceBlockingStub authenticatedStub;
@@ -69,32 +56,20 @@ public class TimeSeriesGrpcInsertResultTypeIT extends BaseGraphServerTest {
   @BeforeEach
   void setupGrpcClient() {
     channel = ManagedChannelBuilder.forAddress("localhost", GRPC_PORT).usePlaintext().build();
-    final Channel authenticatedChannel = ClientInterceptors.intercept(channel, new AuthClientInterceptor());
+    final Channel authenticatedChannel = ClientInterceptors.intercept(channel,
+        new GrpcTestAuthInterceptor("root", DEFAULT_PASSWORD_FOR_TESTS, getDatabaseName()));
     authenticatedStub = ArcadeDbServiceGrpc.newBlockingStub(authenticatedChannel);
   }
 
   @AfterEach
   void shutdownGrpcClient() throws InterruptedException {
-    if (channel != null) {
-      channel.shutdown();
-      channel.awaitTermination(5, TimeUnit.SECONDS);
-    }
-    GlobalConfiguration.SERVER_PLUGINS.setValue("");
-  }
-
-  private final class AuthClientInterceptor implements ClientInterceptor {
-    @Override
-    public <ReqT, RespT> ClientCall<ReqT, RespT> interceptCall(
-        final MethodDescriptor<ReqT, RespT> method, final CallOptions callOptions, final Channel next) {
-      return new ForwardingClientCall.SimpleForwardingClientCall<ReqT, RespT>(next.newCall(method, callOptions)) {
-        @Override
-        public void start(final Listener<RespT> responseListener, final Metadata headers) {
-          headers.put(USER_HEADER, "root");
-          headers.put(PASSWORD_HEADER, DEFAULT_PASSWORD_FOR_TESTS);
-          headers.put(DATABASE_HEADER, getDatabaseName());
-          super.start(responseListener, headers);
-        }
-      };
+    try {
+      if (channel != null) {
+        channel.shutdown();
+        channel.awaitTermination(5, TimeUnit.SECONDS);
+      }
+    } finally {
+      GlobalConfiguration.SERVER_PLUGINS.setValue("");
     }
   }
